rm: FailedEvents, some features from es added here

+++ infrastructure/src/main/kotlin/pl/szymanski/wiktor/ta/infrastructure/scheduler/OfferMakerScheduler.kt
@@ -45,7 +45,7 @@ object OfferMakerScheduler {
             scope.launch {
                 while (isActive) {
                     offerMaker.makeOffers()
-                    delay(config.intervalSeconds * MILLIS_IN_SECOND)
+                    delay((config.intervalSeconds * MILLIS_IN_SECOND).toLong())
                 }
             }
     }
